refactor: move end_user_api_keys endpoints to dedicated api file
